web/divers/models.py
from django.contrib.auth.models import AbstractUser
from django.db import models
from ckeditor.fields import RichTextField

class User(AbstractUser):
    pass

class Post(models.Model):
    content = RichTextField()