web/requirements.txt
django==4.1.2
gunicorn
requests
whitenoise
django-dotenv
psycopg2-binary
django-storages
boto3
django-widget-tweaks
django-crispy-forms
django-ckeditor
django-livereload-server
